+++ CHANGELOG.md
@@ -75,6 +75,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   Macros functionality to find definitions not in c/cpp file
 - Code completions to add documentation to the items to be added where relevant
 - Code Action for node `Property "reg" is required` to add address of node
+- Diagnostic error when a zephyr binding intened for a bus but node is not on bus
 
 ### Fixed